pyproject.toml
[tool.poetry]
name = "voicechat"
version = "0.1.0"
description = ""
authors = ["Tu Nguyen"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.5"
realtimestt = "^0.1.7"
realtimetts = "^0.2.0"
llama-cpp-python = "^0.2.17"
huggingface-hub = "^0.19.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
